Update home page. Make login page work
This commit add a ton of minor details on how everything is laid out.
The login feature works. When correct user is logged in, it redirects
them back to home page. If it doesnt, an error pops up while still being
in the login page. Also, added tailwind configs to find the login.html.
That was giving me major issues at first.